requirements.txt
                                                                                         streamlit
groq
PyPDF2
pdfplumber
pandas
numpy
deep-translator
gTTS
bcrypt
python-dotenv
pymupdf
                                                                                           